refactor(coordinator): rename worker supervision internals
