+++ docs/implementation-status.md
@@ -81,8 +81,8 @@ Single source of truth for the gap between accepted design and running code. Oth
 | On-demand worker manager (spawn on demand, idle shutdown, max active) | ✅ | `packages/coordinator/src/worker-manager/manager.ts` | Used by `apps/server/src/execution/coordinator.ts` |
 | Subagent as bound extension (context-inheriting, ticketless, gate-exempt) | ✅ | `packages/openomni/src/subagent/` | `SubagentRuntime` + `BackgroundManager` — NOT a worker tier; workers are always isolated processes (ADR-010 §6) |
 | Extension-vs-independence routing (subagent vs worker choice in Resident flow) | 📋 | — | ADR-010 §6 — "needs my context, or its own footing?"; domain expertise = independence signal |
-| Worker supervisor internals | ✅ | `packages/coordinator/src/worker-pool/supervisor.ts` | Shared process supervisor used by `worker-manager`; legacy `createWorkerPool` facade removed |
-| SessionRouting helper | ✅ | `packages/coordinator/src/worker-pool/session-routing.ts`, `packages/coordinator/src/worker-manager/manager.ts` | `worker-manager` owns an isolated `SessionRouter` instance for live session-to-slot affinity; the old exported singleton and worker-pool barrel have been removed |
+| Worker supervisor internals | ✅ | `packages/coordinator/src/worker-supervision/supervisor.ts` | Shared process supervisor used by `worker-manager`; legacy `createWorkerPool` facade removed |
+| SessionRouting helper | ✅ | `packages/coordinator/src/worker-supervision/session-routing.ts`, `packages/coordinator/src/worker-manager/manager.ts` | `worker-manager` owns an isolated `SessionRouter` instance for live session-to-slot affinity; the old exported singleton and worker-supervision barrel have been removed |
 | Crash recovery (mark interrupted at boot) | ✅ | `packages/coordinator/src/recovery/` | |
 | Injection queue (async delivery at turn.finish) | ✅ | `packages/openomni/src/execution-runtime/injection-queue.ts` | |
 | Cron job registry persistence | ✅ | `packages/openomni/src/execution-runtime/cron-job-registry.ts`, `packages/session/src/storage/sqlite-cron-job-adapter.ts` | `schedule.create` jobs persist in SQLite and survive `Storage` reinitialize; `schedule.cancel` removes persisted jobs |
